refactor: create reusable components and simplify directories structure
